backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["aiogram (>=3.18.0,<4.0.0)", "pydantic-settings (>=2.8.1,<3.0.0)", "fastapi[uvicorn] (>=0.115.11,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)", "tortoise-orm[asyncpg] (>=0.24.2,<0.25.0)", "asyncpg (>=0.30.0,<0.31.0)"]

[[project.authors]]
name = "Your Name"
email = "you@example.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
